+++ _examples/request-body/read-query/main.go
@@ -25,6 +25,12 @@ func main() {
 		ctx.Writef("MyType: %#v", t)
 	})
 
+	app.Get("/simple", func(ctx iris.Context) {
+		names := ctx.URLParamSlice("name")
+		ctx.Writef("names: %v", ids)
+	})
+
 	// http://localhost:8080?name=iris&age=3
+	// http://localhost:8080/simple?name=john&name=doe&name=kataras
 	app.Listen(":8080")
 }